requirements.txt
flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7
pillow==10.2.0
opencv-python-headless==4.8.1.78
scikit-image==0.21.0
imageio==2.31.5
pyvips==2.2.1
numpy==1.24.3